requirements.txt
streamlit==1.12.0
pandas
snowflake-connector-python[pandas]
altair
toml